Remove the system time channels on Linux targets
A limitation in the current VeriStand support for Action VIs
(that ActionVIOnTargetTypeChange is not implemented) means that
we need to check the target type at other entry points in order
to hide the channels. The result is that the channels will not
(dis)appear in/from the system explorer tree when the target OS
changes, until the custom device is opened or the system definition
is saved.