Default behavior on unsupported OSes should be to fail verbosely at the OS specific bits
Porting to NI Linux RT was tedious due to the lack of granularity in error messages.
Failing in this manner should make porting easier; building and deploying to an unsupported target shows the call stack of the failure.
Alternatively, the user can find callers of Report Unsupported Operating System.vi.
Failure call-stack reporting was tested by changing the OS build code to an unsupported one and attempting to deploy to a Linux setup.